--- a/test/SimpleTransformIterator-test.js
+++ b/test/SimpleTransformIterator-test.js
@@ -1108,10 +1108,6 @@ describe('SimpleTransformIterator', () => {
           result.on('end', done);
         });
 
-        it('should be a SimpleTransformIterator', () => {
-          result.should.be.an.instanceof(SimpleTransformIterator);
-        });
-
         it('should prepend the items', () => {
           items.should.deep.equal(['i', 'ii', 'iii', 'a', 'b', 'c']);
         });
@@ -1136,10 +1132,6 @@ describe('SimpleTransformIterator', () => {
         before(done => {
           result.on('data', item => { items.push(item); });
           result.on('end', done);
-        });
-
-        it('should be a SimpleTransformIterator', () => {
-          result.should.be.an.instanceof(SimpleTransformIterator);
         });
 
         it('should append the items', () => {