supercedes #65

The bolded performance regression needs to be ironed out first.

Before
For loop with 5x10^9 elems: 5.207s
UnionIterator 2x10^7 iterators: 30.297s
**UnionIterator 1000x500 iterators: 1.897s**
UnionIterator 1000x500 iterators - max parallelism of 1: 1.837s

After
For loop with 5x10^9 elems: 6.349s
UnionIterator 2x10^7 iterators: 3.513s
**UnionIterator 1000x500 iterators: 7.219s**
UnionIterator 1000x500 iterators - max parallelism of 1: 168.007ms