@RubenVerborgh 2022-08-22
See https://github.com/RubenVerborgh/AsyncIterator/pull/81